REF: Move some conversions into backend modules.
